+++ src/components/faq/templates/Cmdf2021.jsx
@@ -5,7 +5,7 @@ import {
   Columns,
   Column,
 } from '@lib/Containers'
-import ExpandableFaqList from '@components/faq/shared/FaqList'
+import ExpandableFaqList from '@components/faq/lib/FaqList'
 import backgroundPlants from '@assets/faq__background_plants.svg'
 import titleOptions from '@assets/faq__title_options.svg'
 import questionLeftDecor from '@assets/faq__question_leftdecor.svg'
