+++ src/components/faq/templates/ExpandableWithCategories.jsx
@@ -2,7 +2,7 @@ import React, { useEffect, useState } from 'react'
 import styled from 'styled-components'
 import { SectionContainer, Columns, Column } from '@lib/Containers'
 import Spacers from '@lib/Helpers'
-import ExpandableFaqList from '../shared/FaqList'
+import ExpandableFaqList from '../lib/FaqList'
 
 const TitleImg = styled.img`
   margin-bottom: ${p => p.marginBottom};
